+++ src/o365/graph/commands/teams/teams-message-list.ts
@@ -5,9 +5,9 @@ import GlobalOptions from '../../../../GlobalOptions';
 import {
   CommandOption, CommandValidate
 } from '../../../../Command';
-import GraphCommand from "../../GraphCommand";
-import * as request from 'request-promise-native';
+import { GraphItemsListCommand } from '../GraphItemsListCommand';
 import Utils from '../../../../Utils';
+import { Message } from './Message';
 
 const vorpal: Vorpal = require('../../../../vorpal-init');
 
@@ -18,52 +18,41 @@ interface CommandArgs {
 interface Options extends GlobalOptions {
   teamId: string;
   channelId: string;
-  messageId: string;
 }
 
-class GraphTeamsChannelMessageGetCommand extends GraphCommand {
+class GraphTeamsMessageListCommand extends GraphItemsListCommand<Message> {
   public get name(): string {
-    return `${commands.TEAMS_CHANNEL_MESSAGE_GET}`;
+    return `${commands.TEAMS_MESSAGE_LIST}`;
   }
 
   public get description(): string {
-    return 'Retrieves a message from a channel in a Microsoft Teams team';
+    return 'Lists all messages from a channel in a Microsoft Teams team';
   }
 
   public commandAction(cmd: CommandInstance, args: CommandArgs, cb: () => void): void {
-    auth
-      .ensureAccessToken(auth.service.resource, cmd, this.debug)
-      .then((): request.RequestPromise => {
-        const requestOptions: any = {
-          url: `${auth.service.resource}/beta/teams/${args.options.teamId}/channels/${args.options.channelId}/messages/${args.options.messageId}`,
-          headers: Utils.getRequestHeaders({
-            authorization: `Bearer ${auth.service.accessToken}`,
-            accept: 'application/json;odata.metadata=none'
-          }),
-          json: true
-        };
+    const endpoint: string = `${auth.service.resource}/beta/teams/${args.options.teamId}/channels/${args.options.channelId}/messages`;
 
-        if (this.debug) {
-          cmd.log('Executing web request...');
-          cmd.log(requestOptions);
-          cmd.log('');
+    this
+    .getAllItems(endpoint, cmd, true)
+    .then((): void => {
+        if (args.options.output === 'json') {
+          cmd.log(this.items);
         }
-
-        return request.get(requestOptions);
-      })
-      .then((res: any): void => {
-        if (this.debug) {
-          cmd.log('Response:')
-          cmd.log(res);
-          cmd.log('');
+        else {
+          cmd.log(this.items.map(m => {
+            return {
+              id: m.id,
+              summary: m.summary,
+              body: m.body.content
+            }
+          }));
         }
-
-        cmd.log(res);
+        
         if (this.verbose) {
           cmd.log(vorpal.chalk.green('DONE'));
         }
         cb();
-      }, (err: any): void => this.handleRejectedODataJsonPromise(err, cmd, cb));
+    }, (err: any): void => this.handleRejectedODataJsonPromise(err, cmd, cb));
   }
 
   public options(): CommandOption[] {
@@ -74,11 +63,7 @@ class GraphTeamsChannelMessageGetCommand extends GraphCommand {
       },
       {
         option: '-c, --channelId <channelId>',
-        description: 'The ID of the channel that contains the message'
-      },
-      {
-        option: '-m, --messageId <messageId>',
-        description: 'The ID of the message to retrieve'
+        description: 'The ID of the channel for which to list messages'
       }
     ];
 
@@ -100,10 +85,6 @@ class GraphTeamsChannelMessageGetCommand extends GraphCommand {
         return 'Required parameter channelId missing';
       }
 
-      if (!args.options.messageId) {
-        return 'Required parameter messageId missing';
-      }
-
       return true;
     };
   }
@@ -121,19 +102,19 @@ class GraphTeamsChannelMessageGetCommand extends GraphCommand {
     in preview and is subject to change once the API reached general
     availability.
 
-    To retrieve a message from a Microsoft Teams channel, you have to first log
+    To lists all the messages from a Microsoft Teams channel, you have to first log
     in to the Microsoft Graph using the ${chalk.blue(commands.LOGIN)} command,
     eg. ${chalk.grey(`${config.delimiter} ${commands.LOGIN}`)}.
 
-    You can only retrieve a message from a Microsoft Teams team if you are
+    You can list all the messages from a Microsoft Teams team if you are
     a member of that team.
 
   Examples:
   
-    Retrieve the specified message from a channel of the Microsoft Teams team
-      ${chalk.grey(config.delimiter)} ${this.name} --teamId 5f5d7b71-1161-44d8-bcc1-3da710eb4171 --channelId 19:88f7e66a8dfe42be92db19505ae912a8@thread.skype --messageId 1540747442203
+    Lists all the messages from a channel of the Microsoft Teams team
+      ${chalk.grey(config.delimiter)} ${this.name} --teamId fce9e580-8bba-4638-ab5c-ab40016651e3 --channelId 19:eb30973b42a847a2a1df92d91e37c76a@thread.skype
 `   );
   }
 }
 
-module.exports = new GraphTeamsChannelMessageGetCommand();
+module.exports = new GraphTeamsMessageListCommand();